backend/requirements.txt
# PRD Section 8 — versions pinned when locking for production
fastapi
python-multipart
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv
httpx
beautifulsoup4
pymupdf
google-generativeai
supabase
python-jose[cryptography]
pytest
pytest-asyncio
